requirements.txt
python-telegram-bot==20.*
yt-dlp>=2023.10.13
python-dotenv
